Modernization
This commit modernizes the code to be typescript based and adds lots of
typing for safety. It still requires a lot of cleanup and fixing up of
some of the general monorepo settings but overall it works.

This allows us to now define a protocol that the ObsController returns
as a JSON object. This protocol lists scenes available to the controller
and what actions can be used to modify those scenes. It allows for
providing either options or props. Props are input fields that will
allow you to put in any value such as scaling the scene to say 3.0, or
0.1. Options generate buttons with pre-set values such as a button to
scale the scene to 0.25, or another button for 0.5.