Update OBS State on change
This commit makes it so that as changes occur we update the obs
controller state file so it's a better representation of the overall
state at any given time. This makes it so that we can reload from the
state which helps in development and if for some reason there was a
crash in an actual run we would be able to "resume" on our next change.