Converting to a JS based config file
Taking a page from Vite I've decided to redo the config file as a simple
JS file we will load named `thebit.config.js`. This will make it easier
for anyone making listeners to build a javascript function. It also
makes it a bit easier to pass the function to the VM sandbox for
execution.